Add support for textures and runtime array descriptors
Added support for loading textures from gltf. This also means I want to
have runtime arrays for descriptors to be able to easily handle binding
textures in my path tracer so had to add those too.

Also had to add alignment support for buffer creation because it turns
out so far I just got lucky with the acceleration structure scratch
buffer's alignment but the code wasn't actually correct because it
didn't respect the min alignment rules for the scratch buffer.